Added basic work for Conditionals
I'm still iffy on obtaining the actual value for Compares and Contains;
I've gone as far as fetching the MatchableFields for the Expressions but
need to consult whoever wrote MatchableFields to better understand this.
My instinct was to actually create a table with two rows with a
matchable field for the IF statement, but Lee stated that the
conditional can return a boolean and resolve directly into the
flowtables for then or else. Will clarify with you guys once I get there
at 8 and fix it.